packages/core/pyproject.toml
[project]
name = "clawopc-core"
version = "0.1.0"
description = "ClawOPC core library — workspace, task, role, worklog operations"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawopc_core"]
